[AN] refactor(#377): SP 패키지 변경
common -> data
