+++ android/gitudy/app/build.gradle.kts
@@ -57,7 +57,6 @@ android {
 }
 
 dependencies {
-    implementation(project(":common"))
     implementation(project(":data"))
     implementation(project(":domain"))
     implementation(project(":presentation"))
